+++ share/src/main/java/com/onlyoffice/web/evaluator/IsConvertPermission.java
@@ -8,6 +8,7 @@ package com.onlyoffice.web.evaluator;
 import com.onlyoffice.web.scripts.OnlyofficeSettingsQuery;
 import org.alfresco.error.AlfrescoRuntimeException;
 import org.alfresco.web.evaluator.BaseEvaluator;
+import org.json.simple.JSONArray;
 import org.json.simple.JSONObject;
 
 public class IsConvertPermission extends BaseEvaluator {
@@ -20,33 +21,71 @@ public class IsConvertPermission extends BaseEvaluator {
     @Override
     public boolean evaluate(JSONObject jsonObject) {
         try {
-            if (onlyofficeSettings.getConvertOriginal()) {
-                JSONObject node = (JSONObject)jsonObject.get("node");
-                if (node != null && node.containsKey("permissions")) {
-                    JSONObject perm = (JSONObject)node.get("permissions");
-                    if (perm != null && perm.containsKey("user")) {
-                        JSONObject user = (JSONObject)perm.get("user");
-                        if (user != null && (boolean)user.getOrDefault("Write", false)) {
-                            return true;
-                        }
-                    }
-                }
-            } else {
-                JSONObject parent = (JSONObject)jsonObject.get("parent");
-                if (parent != null && parent.containsKey("permissions")) {
-                    JSONObject perm = (JSONObject)parent.get("permissions");
-                    if (perm != null && perm.containsKey("user")) {
-                        JSONObject user = (JSONObject)perm.get("user");
-                        if (user != null && (boolean)user.getOrDefault("CreateChildren", false)) {
-                            return true;
-                        }
-                    }
-                }
-            }
-
-            return false;
+            return hasPermission(jsonObject) && isConvertibleFormat(jsonObject);
         } catch (Exception err) {
             throw new AlfrescoRuntimeException("Failed to run action evaluator", err);
         }
+    }
+
+    private boolean hasPermission (JSONObject jsonObject) {
+        if (onlyofficeSettings.getConvertOriginal()) {
+            JSONObject node = (JSONObject) jsonObject.get("node");
+            if (node != null && node.containsKey("permissions")) {
+                JSONObject perm = (JSONObject) node.get("permissions");
+                if (perm != null && perm.containsKey("user")) {
+                    JSONObject user = (JSONObject) perm.get("user");
+                    if (user != null && (boolean) user.getOrDefault("Write", false)) {
+                        return true;
+                    }
+                }
+            }
+        } else {
+            JSONObject parent = (JSONObject) jsonObject.get("parent");
+            if (parent != null && parent.containsKey("permissions")) {
+                JSONObject perm = (JSONObject) parent.get("permissions");
+                if (perm != null && perm.containsKey("user")) {
+                    JSONObject user = (JSONObject) perm.get("user");
+                    if (user != null && (boolean) user.getOrDefault("CreateChildren", false)) {
+                        return true;
+                    }
+                }
+            }
+        }
+
+        return false;
+    }
+
+
+    private boolean isConvertibleFormat(JSONObject jsonObject) {
+        String fileName = (String) jsonObject.get("fileName");
+        String docExt = fileName.substring(fileName.lastIndexOf(".") + 1).trim().toLowerCase();
+
+        JSONArray supportedFormats = onlyofficeSettings.getSupportedFormats();
+
+        for (int i = 0; i < supportedFormats.size(); i++) {
+            JSONObject format = (JSONObject) supportedFormats.get(i);
+            JSONArray outputTypes = (JSONArray) format.get("convertTo");
+
+            if (format.get("name").equals(docExt)) {
+                switch (format.get("type").toString()) {
+                    case "FORM":
+                        if (outputTypes.contains("oform")) return true;
+                        break;
+                    case "WORD":
+                        if (outputTypes.contains("docx")) return true;
+                        break;
+                    case "CELL":
+                        if (outputTypes.contains("xlsx")) return true;
+                        break;
+                    case "SLIDE":
+                        if (outputTypes.contains("pptx")) return true;
+                        break;
+                    default:
+                        break;
+                }
+            }
+        }
+
+        return false;
     }
 }